refactor: update rule imports for new model categories
