+++ src/rules/coloring_qubo.rs
@@ -9,7 +9,7 @@
 //! QUBO has n*K variables.
 
 use crate::models::graph::KColoring;
-use crate::models::optimization::QUBO;
+use crate::models::algebraic::QUBO;
 use crate::reduction;
 use crate::rules::traits::{ReduceTo, ReductionResult};
 use crate::topology::{Graph, SimpleGraph};
