+++ src/rules/maximumindependentset_qubo.rs
@@ -6,7 +6,7 @@
 //! Q[i][i] = -w_i, Q[i][j] = P for edges. P = 1 + Σ w_i.
 
 use crate::models::graph::MaximumIndependentSet;
-use crate::models::optimization::QUBO;
+use crate::models::algebraic::QUBO;
 use crate::reduction;
 use crate::rules::traits::{ReduceTo, ReductionResult};
 use crate::topology::{Graph, SimpleGraph};
