+++ src/rules/maximummatching_ilp.rs
@@ -7,7 +7,7 @@
 //! - Objective: Maximize the sum of weights of selected edges
 
 use crate::models::graph::MaximumMatching;
-use crate::models::optimization::{LinearConstraint, ObjectiveSense, VarBounds, ILP};
+use crate::models::algebraic::{LinearConstraint, ObjectiveSense, VarBounds, ILP};
 use crate::reduction;
 use crate::rules::traits::{ReduceTo, ReductionResult};
 use crate::topology::{Graph, SimpleGraph};
