+++ src/rules/maximumsetpacking_ilp.rs
@@ -5,7 +5,7 @@
 //! - Constraints: x_i + x_j <= 1 for each overlapping pair (i, j)
 //! - Objective: Maximize the sum of weights of selected sets
 
-use crate::models::optimization::{LinearConstraint, ObjectiveSense, VarBounds, ILP};
+use crate::models::algebraic::{LinearConstraint, ObjectiveSense, VarBounds, ILP};
 use crate::models::set::MaximumSetPacking;
 use crate::reduction;
 use crate::rules::traits::{ReduceTo, ReductionResult};
