+++ src/rules/minimumdominatingset_ilp.rs
@@ -7,7 +7,7 @@
 //! - Objective: Minimize the sum of weights of selected vertices
 
 use crate::models::graph::MinimumDominatingSet;
-use crate::models::optimization::{LinearConstraint, ObjectiveSense, VarBounds, ILP};
+use crate::models::algebraic::{LinearConstraint, ObjectiveSense, VarBounds, ILP};
 use crate::reduction;
 use crate::rules::traits::{ReduceTo, ReductionResult};
 use crate::topology::{Graph, SimpleGraph};
