+++ src/rules/sat_minimumdominatingset.rs
@@ -15,7 +15,7 @@
 //! - Selecting the dummy vertex means the variable can be either (unused in any clause)
 
 use crate::models::graph::MinimumDominatingSet;
-use crate::models::satisfiability::Satisfiability;
+use crate::models::formula::Satisfiability;
 use crate::reduction;
 use crate::rules::sat_maximumindependentset::BoolVar;
 use crate::rules::traits::{ReduceTo, ReductionResult};
